Safely copy linked dir.

+++ fapistrano/plugins/localshared.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 
-from fabric.api import env, run
+from fabric.api import env, run, exists
 from .. import signal, configuration, directory
 
 def init():
@@ -29,10 +29,13 @@ def _check_shared_dir_parents():
 def _copy_linked_files():
     for linked_file in directory.get_linked_files():
         env.linked_file = linked_file
-        run('cp %(localshared_source)s/%(linked_file)s %(shared_path)s/%(linked_file)s' % env)
+        if exists('%(localshared_source)s/%(linked_file)s' % env):
+            run('cp %(localshared_source)s/%(linked_file)s %(shared_path)s/%(linked_file)s' % env)
         del env['linked_file']
 
 def _copy_linked_dirs():
     for linked_dir in directory.get_linked_dirs():
         env.linked_dir = linked_dir
-        run('cp -R %(localshared_source)s/%(linked_dir)s/* %(shared_path)s/%(linked_dir)s' % env)
+        if exists('%(localshared_source)s/%(linked_dir)s' % env):
+            run('cp -R %(localshared_source)s/%(linked_dir)s/* %(shared_path)s/%(linked_dir)s' % env)
+        del env['linked_dir']
